Updating debugAdapterExecutable to new API
https://code.visualstudio.com/updates/v1_30#_debugging-api has
depricated adapterExecutableCommand. This PR updates the extension to
use DebugAdapterDescriptorFactory.
